+++ gantry/commands/follow.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 import click
-from beaker import Beaker, Experiment
+from beaker import BeakerWorkload
 from rich import print
 
 from .. import util
@@ -10,12 +10,15 @@ from .main import CLICK_COMMAND_DEFAULTS, main
 
 
 @main.command(**CLICK_COMMAND_DEFAULTS)
-@click.argument("experiment", nargs=1, required=False, type=str)
+@click.argument("workload", nargs=1, required=False, type=str)
 @click.option(
     "-t", "--tail", is_flag=True, help="Only tail the logs as opposed to printing all logs so far."
 )
 @click.option(
-    "-l", "--latest", is_flag=True, help="Get the logs from the latest running experiment."
+    "-l",
+    "--latest",
+    is_flag=True,
+    help="Get the logs from the latest running experiment (non-session) workload.",
 )
 @click.option(
     "-w",
@@ -28,10 +31,10 @@ from .main import CLICK_COMMAND_DEFAULTS, main
     "-a",
     "--author",
     type=str,
-    help="Pull the latest experiment for a particular author. Defaults to your own account.",
+    help="Pull the latest experiment workload for a particular author. Defaults to your own account.",
 )
 def follow(
-    experiment: Optional[str] = None,
+    workload: Optional[str] = None,
     tail: bool = False,
     latest: bool = False,
     workspace: Optional[str] = None,
@@ -40,29 +43,30 @@ def follow(
     """
     Follow the logs for a running experiment.
     """
-    beaker = Beaker.from_env(session=True)
+    with util.init_client(ensure_workspace=False) as beaker:
+        wl: Optional[BeakerWorkload] = None
+        if workload is not None:
+            if latest or workspace is not None or author is not None:
+                raise ConfigurationError(
+                    "[WORKLOAD] is mutually exclusive with -a/--author, -w/--workspace, and -l/--latest"
+                )
+            wl = beaker.workload.get(workload)
+        else:
+            if not latest:
+                raise ConfigurationError(
+                    "A filter such as -l/--latest is required when no [WORKLOAD] is specified"
+                )
 
-    exp: Optional[Experiment] = None
-    if experiment is not None:
-        if latest or workspace is not None or author is not None:
-            raise ConfigurationError(
-                "EXPERIMENT is mutually exclusive with -a/--author, -w/--workspace, and -l/--latest"
+            wl = util.get_latest_workload(
+                beaker, author_name=author, workspace_name=workspace, running=True
             )
-        exp = beaker.experiment.get(experiment)
-    else:
-        if not latest:
-            raise ConfigurationError(
-                "A filter such as -l/--latest is required when no [EXPERIMENT] is specified"
+            if wl is None:
+                raise NotFoundError("Failed to find an experiment workload to follow")
+
+            print(
+                f"Following experiment [b cyan]{wl.experiment.name}[/] ({wl.experiment.id}) at {beaker.workload.url(wl)}"
             )
 
-        exp = util.get_latest_experiment(beaker, author=author, workspace=workspace, running=True)
-        if exp is None:
-            raise NotFoundError("Failed to find an experiment to follow")
-
-        print(
-            f"Following experiment [b cyan]{exp.name}[/] ({exp.id}) at {beaker.experiment.url(exp)}"
-        )
-
-    assert exp is not None
-    job = util.follow_experiment(beaker, exp, tail=tail)
-    util.display_results(beaker, exp, job)
+        assert wl is not None
+        job = util.follow_workload(beaker, wl, tail=tail)
+        util.display_results(beaker, wl, job)